Tuning getters and setters
Refactoring into separate files
