+++ Cargo.toml
@@ -24,6 +24,7 @@ thiserror = "2"
 
 nom = "8"
 derive_builder = "0.20"
+getset = "0.1.6"
 
 [dev-dependencies]
 env_logger = "0.11"
