+++ src/provider/mod.rs
@@ -1,0 +1,3 @@
+// generic
+// gitlab (subgroups) style
+// azure devops
