+++ tests/parse.rs
@@ -4,11 +4,11 @@ fn ssh_user_ports() {
     let test_url = "ssh://git@host.tld:9999/user/project-name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Ssh))
-        .host(Some(String::from("host.tld")))
-        .user(Some(String::from("git")))
-        .port(Some(9999))
-        .path(String::from("user/project-name.git"))
+        .scheme(Scheme::Ssh)
+        .host("host.tld")
+        .user("git")
+        .port(9999 as u16)
+        .path("user/project-name.git")
         .print_scheme(true)
         .build()
         .unwrap();
@@ -21,10 +21,9 @@ fn ssh_no_scheme_no_user() {
     let test_url = "host.tld:user/project-name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Ssh))
-        .host(Some(String::from("host.tld")))
-        .path(String::from("user/project-name.git"))
-        .print_scheme(false)
+        .scheme(Scheme::Ssh)
+        .host("host.tld")
+        .path("user/project-name.git")
         .build()
         .unwrap();
 
@@ -37,10 +36,10 @@ fn https_user_bitbucket() {
     let test_url = "https://user@bitbucket.org/user/repo.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Https))
-        .host(Some(String::from("bitbucket.org")))
-        .user(Some(String::from("user")))
-        .path(String::from("/user/repo.git"))
+        .scheme(Scheme::Https)
+        .host("bitbucket.org")
+        .user("user")
+        .path("/user/repo.git")
         .print_scheme(true)
         .build()
         .unwrap();
@@ -53,10 +52,10 @@ fn ssh_user_bitbucket() {
     let test_url = "git@bitbucket.org:user/repo.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .host(Some(String::from("bitbucket.org")))
-        .scheme(Some(Scheme::Ssh))
-        .user(Some(String::from("git")))
-        .path(String::from("user/repo.git"))
+        .scheme(Scheme::Ssh)
+        .host("bitbucket.org")
+        .user("git")
+        .path("user/repo.git")
         .build()
         .unwrap();
 
@@ -68,11 +67,11 @@ fn https_user_auth_bitbucket() {
     let test_url = "https://x-token-auth:token@bitbucket.org/owner/name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Https))
-        .host(Some("bitbucket.org".to_string()))
-        .user(String::from("x-token-auth"))
-        .token(String::from("token"))
-        .path(String::from("/owner/name.git"))
+        .scheme(Scheme::Https)
+        .host("bitbucket.org")
+        .user("x-token-auth")
+        .token("token")
+        .path("/owner/name.git")
         .print_scheme(true)
         .build()
         .unwrap();
@@ -85,10 +84,10 @@ fn https_user_github() {
     let test_url = "https://user@github.com/user/repo.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Https))
-        .user(Some(String::from("user")))
-        .host(Some(String::from("github.com")))
-        .path(String::from("/user/repo.git"))
+        .scheme(Scheme::Https)
+        .user("user")
+        .host("github.com")
+        .path("/user/repo.git")
         .print_scheme(true)
         .build()
         .unwrap();
@@ -101,10 +100,10 @@ fn ssh_user_github() {
     let test_url = "git@github.com:user/repo.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Ssh))
-        .user(Some(String::from("git")))
-        .host(Some(String::from("github.com")))
-        .path(String::from("user/repo.git"))
+        .scheme(Scheme::Ssh)
+        .user("git")
+        .host("github.com")
+        .path("user/repo.git")
         .build()
         .unwrap();
 
@@ -116,11 +115,11 @@ fn https_user_auth_github() {
     let test_url = "https://token:x-oauth-basic@github.com/owner/name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Https))
-        .user(Some(String::from("token")))
-        .token(Some(String::from("x-oauth-basic")))
-        .host(Some(String::from("github.com")))
-        .path(String::from("/owner/name.git"))
+        .scheme(Scheme::Https)
+        .user("token")
+        .token("x-oauth-basic")
+        .host("github.com")
+        .path("/owner/name.git")
         .print_scheme(true)
         .build()
         .unwrap();
@@ -133,10 +132,10 @@ fn ssh_user_azure_devops() {
     let test_url = "git@ssh.dev.azure.com:v3/CompanyName/ProjectName/RepoName";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Ssh))
-        .user(Some(String::from("git")))
-        .host(Some(String::from("ssh.dev.azure.com")))
-        .path(String::from("v3/CompanyName/ProjectName/RepoName"))
+        .scheme(Scheme::Ssh)
+        .user("git")
+        .host("ssh.dev.azure.com")
+        .path("v3/CompanyName/ProjectName/RepoName")
         .build()
         .unwrap();
 
@@ -148,10 +147,10 @@ fn https_user_azure_devops() {
     let test_url = "https://organization@dev.azure.com/organization/project/_git/repo";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Https))
-        .user(Some(String::from("organization")))
-        .host(Some(String::from("dev.azure.com")))
-        .path(String::from("/organization/project/_git/repo"))
+        .scheme(Scheme::Https)
+        .user("organization")
+        .host("dev.azure.com")
+        .path("/organization/project/_git/repo")
         .print_scheme(true)
         .build()
         .unwrap();
@@ -164,10 +163,10 @@ fn ftp_user() {
     let test_url = "ftp://git@host.tld/user/project-name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Ftp))
-        .user(Some(String::from("git")))
-        .host(Some(String::from("host.tld")))
-        .path(String::from("/user/project-name.git"))
+        .scheme(Scheme::Ftp)
+        .user("git")
+        .host("host.tld")
+        .path("/user/project-name.git")
         .print_scheme(true)
         .build()
         .unwrap();
@@ -180,10 +179,10 @@ fn ftps_user() {
     let test_url = "ftps://git@host.tld/user/project-name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Ftps))
-        .user(Some(String::from("git")))
-        .host(Some(String::from("host.tld")))
-        .path(String::from("/user/project-name.git"))
+        .scheme(Scheme::Ftps)
+        .user("git")
+        .host("host.tld")
+        .path("/user/project-name.git")
         .print_scheme(true)
         .build()
         .unwrap();
@@ -196,8 +195,8 @@ fn relative_unix_path() {
     let test_url = "../project-name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::File))
-        .path(String::from("../project-name.git"))
+        .scheme(Scheme::File)
+        .path("../project-name.git")
         .build()
         .unwrap();
 
@@ -209,8 +208,8 @@ fn absolute_unix_path() {
     let test_url = "/path/to/project-name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::File))
-        .path(String::from("/path/to/project-name.git"))
+        .scheme(Scheme::File)
+        .path("/path/to/project-name.git")
         .build()
         .unwrap();
 
@@ -223,8 +222,8 @@ fn relative_windows_path() {
     let test_url = "..\\project-name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::File))
-        .path(String::from("../project-name.git"))
+        .scheme(Scheme::File)
+        .path("../project-name.git")
         .build()
         .unwrap();
 
@@ -239,8 +238,8 @@ fn absolute_windows_path() {
     let test_url = "c:\\project-name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::File))
-        .path(String::from("c:\\project-name.git"))
+        .scheme(Scheme::File)
+        .path("c:\\project-name.git")
         .build()
         .unwrap();
 
@@ -310,9 +309,9 @@ fn git() {
     let test_url = "git://github.com/owner/name.git";
     let parsed = GitUrl::parse(test_url).expect("URL parse failed");
     let expected = GitUrlBuilder::default()
-        .scheme(Some(Scheme::Git))
-        .host(Some(String::from("github.com")))
-        .path(String::from("/owner/name.git"))
+        .scheme(Scheme::Git)
+        .host("github.com")
+        .path("/owner/name.git")
         .print_scheme(true)
         .build()
         .unwrap();
